+++ applications/archive/views/archive_main_view.h
@@ -26,36 +26,23 @@ typedef enum {
     ArchiveTabTotal,
 } ArchiveTabEnum;
 
-static const char* known_ext[] = {
-    [ArchiveFileTypeIButton] = ".ibtn",
-    [ArchiveFileTypeNFC] = ".nfc",
-    [ArchiveFileTypeSubGhz] = ".sub",
-    [ArchiveFileTypeLFRFID] = ".rfid",
-    [ArchiveFileTypeIrda] = ".ir",
-};
-
-static inline const char* get_tab_ext(ArchiveTabEnum tab) {
-    switch(tab) {
-    case ArchiveTabIButton:
-        return known_ext[ArchiveFileTypeIButton];
-    case ArchiveTabNFC:
-        return known_ext[ArchiveFileTypeNFC];
-    case ArchiveTabSubGhz:
-        return known_ext[ArchiveFileTypeSubGhz];
-    case ArchiveTabLFRFID:
-        return known_ext[ArchiveFileTypeLFRFID];
-    case ArchiveTabIrda:
-        return known_ext[ArchiveFileTypeIrda];
-    default:
-        return "*";
-    }
-}
-
 typedef enum {
-    ArchiveBrowserEventRename,
+    ArchiveBrowserEventFileMenuOpen,
+    ArchiveBrowserEventFileMenuClose,
+    ArchiveBrowserEventFileMenuRun,
+    ArchiveBrowserEventFileMenuPin,
+    ArchiveBrowserEventFileMenuRename,
+    ArchiveBrowserEventFileMenuDelete,
+    ArchiveBrowserEventEnterDir,
     ArchiveBrowserEventExit,
-    ArchiveBrowserEventLeaveDir,
 } ArchiveBrowserEvent;
+
+static const uint8_t file_menu_actions[MENU_ITEMS] = {
+    [0] = ArchiveBrowserEventFileMenuRun,
+    [1] = ArchiveBrowserEventFileMenuPin,
+    [2] = ArchiveBrowserEventFileMenuRename,
+    [3] = ArchiveBrowserEventFileMenuDelete,
+};
 
 typedef struct ArchiveMainView ArchiveMainView;
 
@@ -78,11 +65,11 @@ struct ArchiveMainView {
 
 typedef struct {
     ArchiveTabEnum tab_idx;
-    BrowserActionEnum action;
     files_array_t files;
 
     uint8_t depth;
     uint8_t menu_idx;
+    bool menu;
 
     uint16_t idx;
     uint16_t last_idx[MAX_DEPTH];
@@ -99,21 +86,3 @@ View* archive_main_get_view(ArchiveMainView* main_view);
 
 ArchiveMainView* main_view_alloc();
 void main_view_free(ArchiveMainView* main_view);
-
-void archive_file_array_remove_selected(ArchiveMainView* main_view);
-void archive_file_array_clean(ArchiveMainView* main_view);
-
-void archive_view_add_item(ArchiveMainView* main_view, FileInfo* file_info, const char* name);
-void archive_browser_update(ArchiveMainView* main_view);
-void archive_switch_tab(ArchiveMainView* main_view, InputKey key);
-
-size_t archive_file_array_size(ArchiveMainView* main_view);
-ArchiveFile_t* archive_get_current_file(ArchiveMainView* main_view);
-const char* archive_get_path(ArchiveMainView* main_view);
-const char* archive_get_name(ArchiveMainView* main_view);
-uint8_t archive_get_depth(ArchiveMainView* main_view);
-void archive_set_name(ArchiveMainView* main_view, const char* name);
-
-static inline bool is_known_app(ArchiveFileTypeEnum type) {
-    return (type != ArchiveFileTypeFolder && type != ArchiveFileTypeUnknown);
-}
